Refactored to remove Message class and send strings instead
Protocol for what messages mean is defined in Protocol class.
ObjectStreamWriter/Reader replaced with PrintWriter and BufferedReader.
Moved message verification and authentication to ChatServer class
(previously was split between ChatServer and ServerConnectionThread) to
be more consistent.
Added functionality for "actions" denoted by "/me " at the beginning of
a message.
Made a new "not authenticated" message when a user fails to authenticate
correctly.