removed synchronised keyword, fixed action formatting.
I believe I was using the synchronised keyword incorrectly so I have
removed it for now. I'll read up on how to use it correctly and add it
back in if appropriate.

Actions were formatted on the sender as "Me: /me does whatever" and on
the receiver as "User /me does whatever"