Added private chat functionality, changed public/private protocol
When clicking on a user on the right, a new tab now opens for chatting
directly to that user. The client checks what tab is active before
sending a message and uses SERVER or the tab name in the appropriate
point in the messaging protocol to the server. Server also now handles
as appropriate with new protocol.

ChatFlow class changed to ChatFlowTab and now extends Tab instead of
ScrollPane (contains a ScrollPane as its content).

Added SERVER as a constant in Protocol.

Fixed IndexOutOfBoundsException on server when users exited; USER_EXIT
now is distinct in the switch from MESSAGE and ACTION